+++ kslurm/cli/kapp/__init__.py
@@ -1,0 +1,6 @@
+__submodules__ = []
+
+# <AUTOGEN_INIT>
+__all__ = []
+
+# </AUTOGEN_INIT>
