+++ kslurm/container.py
@@ -1,14 +1,18 @@
 from __future__ import absolute_import
 
+import os
 import platform
 import re
 from pathlib import Path
-from typing import Optional, Union
+from typing import Any, Optional, Union
 
 import attrs
 import requests
 
+from kslurm.appconfig import Config
 from kslurm.args.command import CommandError
+from kslurm.exceptions import ValidationError
+from kslurm.models import validators
 from kslurm.utils import get_hash
 
 REGISTRYBASE = "https://registry-1.docker.io"
@@ -114,7 +118,7 @@ class Container:
         )
 
     def __str__(self):
-        return self.address
+        return self.uri
 
 
 class ManifestError(CommandError):
@@ -182,6 +186,171 @@ def get_digest(container: Container, token: Optional[str] = None) -> Optional[st
     #     # fallback
     #     pass
     print(manifest_json)
+
+
+class SingularityDirError(CommandError):
+    pass
+
+
+class SingularityDir(type(Path())):
+    def __new__(cls):
+        pipdir = Path(Config()["pipdir"])
+        pipdir.mkdir(exist_ok=True)
+        singularity_dir = pipdir / "containers"
+        singularity_dir.mkdir(exist_ok=True)
+        return super().__new__(cls, singularity_dir)
+
+    def __init__(self, *args: Any, **kwargs: Any):
+        super().__init__(*args, **kwargs)
+        self.work.mkdir(exist_ok=True)
+        self.images.mkdir(exist_ok=True)
+        self.uris.mkdir(exist_ok=True)
+        self.snakemake.mkdir(exist_ok=True)
+        self.aliases.mkdir(exist_ok=True)
+
+    @property
+    def work(self):
+        return self / "work"
+
+    @property
+    def images(self):
+        return self / "images"
+
+    @property
+    def uris(self):
+        return self / "uris"
+
+    @property
+    def snakemake(self):
+        return self / "snakemake"
+
+    @property
+    def aliases(self):
+        return self / "aliases"
+
+    def get_data_path(self, container: Container):
+        if container.cache_path:
+            return self.images / container.cache_path
+        result = self.uris / container.uri_path
+        result.parent.mkdir(parents=True, exist_ok=True)
+        return result
+
+    def has_container(self, container: Container):
+        if container.cache_path:
+            return (self.images / container.cache_path).exists()
+        return (self.uris / container.uri_path).exists()
+
+    def update_uri_link(self, container: Container):
+        if container.cache_path:
+            if (cache_path := self.images / container.cache_path).exists():
+                uri_path = self.uris / container.uri_path
+                if uri_path.exists():
+                    if not uri_path.is_symlink():
+                        os.remove(uri_path)
+                    elif cache_path.samefile(os.readlink(uri_path)):
+                        return False
+                    else:
+                        os.unlink(uri_path)
+                uri_path.symlink_to(cache_path)
+                return True
+            raise ValueError()
+        return False
+
+    def find(self, uri_or_alias: str):
+        try:
+            validators.fs_name(uri_or_alias)
+        except ValidationError:
+            container = None
+        else:
+            # Is it in the alias directory
+            if (self.aliases / uri_or_alias).exists():
+                container = ContainerAlias(uri_or_alias, self).image
+            else:
+                container = None
+
+        if container is None:
+            # Does it look like a uri
+            try:
+                container = Container.from_uri(uri_or_alias)
+            except ValueError:
+                raise SingularityDirError(
+                    f"Invalid identifier '{uri_or_alias}'. Must be either an alias for "
+                    "a container, or a valid container uri: "
+                    f"'{Container.URI_SCHEME}'"
+                )
+
+        if (self.uris / container.uri_path).exists():
+            return container
+        return None
+
+    def iter_images(self):
+        for dirpath, _, filenames in os.walk(self.uris):
+            for filename in filenames:
+                uri_path = Path(dirpath, filename).relative_to(self.uris)
+                yield Container.from_uri_path(uri_path)
+
+
+class AliasError(CommandError):
+    pass
+
+
+@attrs.define
+class ContainerAlias:
+    alias: str
+    singularity_dir: SingularityDir
+    _image: Optional[Container] = None
+
+    def __attrs_post_init__(self):
+        self.path.parent.mkdir(exist_ok=True)
+        if self.path.exists():
+            if not self.path.is_symlink():
+                raise AliasError(
+                    f"File or directory found at alias path '{self.path}'. Please "
+                    "manually move or remove this item."
+                )
+            try:
+                self.image
+            except ValueError:
+                raise AliasError(
+                    f"Symlink found at '{self.path}' that does not currently point to "
+                    "a kapp-controlled image. Please manually move or remove this item."
+                )
+
+    @property
+    def path(self):
+        return self.singularity_dir.aliases / self.alias
+
+    @property
+    def image(self):
+        if self._image is None:
+            self._image = Container.from_uri_path(
+                Path(os.readlink(self.path)).relative_to(
+                    self.singularity_dir / "images"
+                )
+            )
+        return self._image
+
+    def __bool__(self) -> bool:
+        return self.path.exists()
+
+    def check_upgrade(self, new_app: Container):
+        if not self.image.is_same_container(new_app):
+            raise AliasError(
+                f"Alias '{self.alias}' is currently pointing to '{self.image.uri}'. To "
+                "change this alias, run again with -f"
+            )
+
+    def link(self, app: Container):
+        if self.path.exists():
+            if app == self.image:
+                print(f"Alias '{self.alias}' already set to '{app}'")
+                return
+            print(f"Updating alias '{self.alias}' from {self.image} to {app}")
+            self.path.unlink()
+        else:
+            print(f"Aliasing {app} as {self.alias}")
+        self.path.symlink_to(self.singularity_dir.images / app.uri_path)
+        self._image = None
 
 
 if __name__ == "__main__":
